Improve --key-direction support for TLS-auth inline files
Creating inline: If --key-direction is specified on the command line
then verify 0|1 and use it.  Otherwise, do not add key-direction to
the inline file.

Renewing inline: If --key-direction is specified on the command line
then verify 0|1 and also verify if this matches the current inline
setting and use it.  If it does not match current inline then issue
a warning.

Confirm renewing before making any changes.

If the inline file will be generated without key-direction then
issue a warning that it is missing.

Closes: https://github.com/TinCanTech/easy-tls/issues/13

Signed-off-by: Richard Bonhomme <tincanteksup@gmail.com>