Inter-active Menu: Add configuring easytls-verify.sh (--tls-verify)
This allows Easy-TLS to completely support an OpenVPN server which
is configured to support both --tla-auth OR --tls-crypt while also
supporting --tls-crypt-v2 keys.

When a client connects without a TLS-Crypt-V2 key then OpenVPN does
not call --tls-crypt-v2-verify.  In order to use the Easy-TLS client
connect script to verify hardware address easytls-verify.sh builds
a fake hardware address file with 000000000000 as the address.
This allows the client to connect as if it has a TLS-Crypt-V2 key
which is not locked to a hardware address.

Note: Openvpn only calls --tls-verify for clients which are configured
with a --tls-auth or --tls-crypt key.

Signed-off-by: Richard Bonhomme <tincanteksup@gmail.com>