Only write a hwaddr file if one does not already exist
If the client is using a --tls-crypt-v2 key then a hwaddr file
will exist, so do not over-write it.

If the client is using either --tls-auth or --tls-crypt
then there will not be a hwaddr file, so create a simple file.
This file implies that the client key is not bound to a hwaddr.

Signed-off-by: Richard Bonhomme <tincanteksup@gmail.com>