Replace 'append' (>>) operator with full 'cat' file concatenation
For what-ever reason, Github CI Windows runner does not append to file
when using '>>' operator.

Instead, it over-writes the portion of the original file from the start,
to the length which corresponds to the length of the appending file.

* Also, replace some failing regular expressions used by grep.

These problems may be due to sh.exe interaction with the host OS.

Signed-off-by: Richard Bonhomme <tincanteksup@gmail.com>