Introduce wait-gate; Drop PID-file and improve Windows integration
* Wait-gate:

It is possible for two clients to try to connect at the same time
with the same X509 Certificate Serial number.  This means that the
client-metadata-file for different clients could have the same name:
${tls_serial_hex_0}.$(EASYTLS_server_pid}

In case this happens the script will now take the following actions:
1. If the old file is more than 60 seconds old then it will be deleted.
2. Otherwise, the script will wait for 2 seconds for the file to be
   deleted by the client-connect script completing.
3. If the file is not deleted then the new client will fail to connect.
4. The new client will try to reconnect for ~60 seconds and then the
   colliding file will be deleted and the new client will be allowed
   to connect.

* Utilise *nix $PPID instead of Openvpn --writepid

This is standard practice for *nix but requires dropping the use of
interim batch files in Openvpn configuration files for Windows.
This means that, except for PATH, the Openvpn configuration files
can be identical between *nix and Windows.

* Windows integration:

Better binary dependency support, inluding $PATH.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>