Complete re-write of easytls-verify.sh
The script now uses the generic_metadata_file to compare states:

If these files exist then:
1. Move generic md file to extended md file and process metadata
   Read metadata into g_* variables
2. If generic md file does not exist and extended md file does exist
   then this must be a renegotiation
3. If there are no generic md files then assume --tls-auth/crypt(v1)
All of these always succeed.

The client X509 serial named md file is then processed:
1. Move client md file to extended md file and process metadata
   Read metadata into c_* variables
2. If client md file does not exist and extended md file does exist
   then this must be a renegotiation or --tls-auth/crypt(v1)
   This is resolved here.
3. Finish processing
Some of these are designed to fail.

Currently, this cannot identify a client which floats but
it does allow floating to occur, which it probably should not.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>