Major rewrite of easytls-verify.sh
I found that with the new vars file, the behaviour of this script
became unpredictable when TLS-Crypt-V2 keys were moved from one
X509 certificate to another.

This patch aims to address the various ways that users might share
there inline files without explicit permission to do so.

The script has also been simplified and standardised, somewhat,
in an attempt to aid future development.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>