Make conntrac errors FATAL
FATAL erros can be configured to kill the server instance.
This is disabled by default.

The source of the problems with easytls-conntrac.lib is an OpenVPN bug:
https://community.openvpn.net/openvpn/ticket/160

When using `--username-as-common-name`, openvpn does not provide
the username to easytls-client-diconnect.sh, instead the certificate
common_name is used.

Using a patched version of openvpn, easytls-conntrac.lib behaves correctly.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>