Use force_hash to generate inline hash during remove-metadata
When using '-y' to disable file hash verification, remove-metadata
did not generate a usable hash of the inline file because it returned
the fixed hash ('2'x64 chars).

Insert 'force_hash=1' (and 'unset force_hash') to generate a valid
inline file hash because a valid hash is required for an inline file.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>