Introduce metadata-file stacking and finally clear all temp-files
There is a small but crucial logical change made to how the type of
TLS key is in use for this client.  This allows for no fake metadata
files when a non-TLS-Crypt-V2 key is used by a client. This change
finally, successfully clears all temp-metadata files on connection.

Also, improve file locking  and tlskey_status()

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>