Add stack reset fail-safe to client-disconnect
Under certain circumstances, the server does not clear metadata files.
This causes the metadata stack to infinitely grow ..
To mitigate this problem, use client-disconnect to clear the stack of
any old (+240 seconds) metadata-temp files.

The circumstances appear to be when the server initiates a renegotiate
and a client is still stuck in deferred-auth.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>