Prepare functions for `faster_hash`
The current master hash (a.k.a. wiscii_hash) is deep and slow.
Each file to be hashed is done so individually.

The new `faster_hash` is a hash of all of the same files as
`wiscii_hash` but piped through a single hash function only once.

This patch allows `faster_hash` functions to be tested.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>