Add variable 'special_errors' for peculiar unit-test problems
The reason for this variable is to catch an error which only occurs
when using the New generate_master_hash() method.

The error is a master-hash mismatch error, during unit-test cross-
pollinated Easy-TLS PKI, when disabling a client TLS-Crypt-V2 key.

The error is Not present when using the Old generate_master_hash()
method.

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>