Introduce TLS-Crypt-V2 Server Key Serial Numbers
This serialisation is only used internally by 'easytls' script,
to keep track of of TLS-Crypt-V2 Server keys within indexes.

The serial-number is a minor modification of the HASH_ALGO of the key file.
SHA256 or SHA1 hash, drop the last four digits and insert an identifyer at
the beginning of the string.

EG:

Standard server:
* SRV:520e023e3278fb7e41d5a8558f8796e09c96e6b68001b22cf753b359ab0e

Group server:
* G_S:98cef97e2fd0639efd291f82583e636fbc9150ceb4e56fa43477c1b6ceb0

SHA1:
* SRV:e36a610abb9560a50f46215bdc5f6a5cf280

Signed-off-by: Richard T Bonhomme <tincantech@protonmail.com>